Fix. Firewall. Refactoring logging process.
