+++ inc/spbc-tools.php
@@ -25,7 +25,7 @@ function spbc_report_country_part($ips_c, $ip = null)
 
 function spbc_report_tc_requests_per($ip = null, $status = null)
 {
-    global $wpdb;
+    global $wpdb, $spbc;
 
     if (is_null($ip) || is_null($status)) {
         return '-';
@@ -49,7 +49,12 @@ function spbc_report_tc_requests_per($ip = null, $status = null)
     );
 
     if (isset($c[0]) && isset($c[0]['entries'])) {
-        return (string)$c[0]['entries'];
+        $entries = (int)$c[0]['entries'];
+        if ($entries > $spbc->settings['traffic_control__autoblock_amount']) {
+            $entries = $spbc->settings['traffic_control__autoblock_amount'];
+        }
+
+        return (string)$entries;
     }
 
     return '-';
