+++ lib/CleantalkSP/SpbctWP/Firewall/BFP.php
@@ -66,7 +66,7 @@ class BFP extends FirewallModule
                             'status' => 'DENY_BY_BFP',
                         )
                     );
-                    FirewallState::$is_need_to_increment_entire = false;
+                    FirewallState::setIsNeedToIncrementEntire(false);
                 }
             }
         }
@@ -141,7 +141,7 @@ class BFP extends FirewallModule
         }
 
         self::$is_checked = true;
-        FirewallState::$is_need_to_increment_entire = false;
+        FirewallState::setIsNeedToIncrementEntire(false);
     }
 
     private function clearTable()
