+++ lib/CleantalkSP/SpbctWP/Firewall/WafBlocker.php
@@ -92,7 +92,7 @@ class WafBlocker extends FirewallModule
         }
 
         $this->is_checked = true;
-        FirewallState::$is_need_to_increment_entire = false;
+        FirewallState::setIsNeedToIncrementEntire(false);
     }
 
     public function clearTable()
