+++ tests/Mock/FirewallStateMock.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace CleantalkSP\SpbctWP\Firewall;
+namespace Mock;
 
-class FirewallState
+class FirewallStateMock
 {
     public static $is_need_to_increment_entire = true;
 
     public static function setIsNeedToIncrementEntire($value)
     {
-        self::$is_need_to_increment_entire = $value;
+        self::$is_need_to_increment_entire = true;
     }
 }
