+++ tests/Standalone/TestBFP.php
@@ -2,6 +2,8 @@
 
 use CleantalkSP\SpbctWP\DB;
 use CleantalkSP\SpbctWP\Firewall\BFP;
+use CleantalkSP\SpbctWP\Firewall\FirewallState;
+use Mock\FirewallStateMock;
 use CleantalkSP\Updater\Updater;
 
 class TestBFP extends PHPUnit\Framework\TestCase
@@ -31,6 +33,8 @@ class TestBFP extends PHPUnit\Framework\TestCase
 
     public function testIpv4Block()
     {
+        class_alias(FirewallStateMock::class, FirewallState::class);
+
         $ip = '127.0.0.1';
         $this->db->execute('
             TRUNCATE TABLE 
@@ -72,5 +76,4 @@ class TestBFP extends PHPUnit\Framework\TestCase
         $this->assertEquals($bfp_result[0]->ip, $ip);
         $this->assertEquals($bfp_result[0]->status, 'DENY_BY_BFP');
     }
-
 }
